ADD :: login refactoring
