+++ src/pages/auth/login/Login.tsx
@@ -1,64 +1,23 @@
 import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { useQuery, useMutation } from "react-query";
-import { getGoogleAuthLink, loginUser } from "../../../apis/auth";
 import { LOGIN_AUTH } from "../../../types/auth.type";
-import { useSetRecoilState } from "recoil";
-import { userState } from "../../../atoms/user";
-import {
-  ACCESS_KEY,
-  REFRESH_KEY,
-  AUTHORITY,
-  NAME,
-} from "../../../constants/user/auth.constant";
-import { GOOGLE_AUTH_LINK } from "../../../constants/keys/auth.keys";
 import * as S from "./Login.style";
-import { Storage } from "../../../lib/storage/storage";
+import {
+  GetGoogleAuthLink,
+  LoginFeature,
+} from "../../../features/auth/login.feature";
 
 export const Login = () => {
-  const navigate = useNavigate();
-  const setUser = useSetRecoilState(userState);
-  const { data } = useQuery([GOOGLE_AUTH_LINK], getGoogleAuthLink);
   const [request, setRequest] = useState<LOGIN_AUTH>({
     email: "",
     password: "",
   });
+  const { login } = LoginFeature(request);
+  const { data } = GetGoogleAuthLink();
 
-  const { mutate } = useMutation(loginUser, {
-    onSuccess: (data) => {
-      Storage.setItem(ACCESS_KEY, data.accessToken);
-      Storage.setItem(REFRESH_KEY, data.refreshToken);
-      // getUser api 만들면 useUSer 만들겠습니다
-      localStorage.setItem(AUTHORITY, data.authority);
-      localStorage.setItem(NAME, data.name);
-      setUser({
-        accessToken: data.accessToken,
-        refreshToken: data.refreshToken,
-        authority: data.authority,
-        name: data.name,
-      });
-
-      if (!data?.login) {
-        navigate("/signup");
-      } else {
-        navigate("/");
-      }
-    },
-  });
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setRequest({
       ...request,
       [e.target.name]: e.target.value,
-    });
-  };
-
-  const login = () => {
-    if (!request.email.endsWith("@bssm.hs.kr")) {
-      request.email += "@bssm.hs.kr";
-    }
-    mutate({
-      email: request.email,
-      password: request.password,
     });
   };
